Remove some warnings and errors
Fix some problems with Linux-based compilation (some missing include files and size_t/int discrepancies)

+++ src/gen.cpp
@@ -151,7 +151,7 @@ int main( int argc, char **argv )
 
 	GridType gt = getGridType( argc, argv );
 
-	size_t idx = 1;
+	int idx = 1;
 
 	while( idx < argc ) {
 		if( !strcmp( argv[idx], "-size" ) ) {
